pyproject.toml
[project]
name = "enter-name-here"
version = "0.2.0"
description = ""
authors = []
requires-python = "~=3.11.12"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.119.0",
    "google-auth>=2.43.0",
    "google-genai>=1.51.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "setuptools>=80.9.0",
    "sqlmodel>=0.0.27",
    "tomlkit>=0.13.3",
    "typer>=0.19.2",
    "uvicorn[standart]>=0.37.0",
]

[tool.pylint.main]
disable = [
    "missing-timeout",
    "missing-module-docstring",
    "missing-class-docstring"
]
extension-pkg-whitelist = ["pydantic"]
ignore-paths = ["^app/database/schemas/*/.*$"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.design]
max-args = 10
max-attributes = 12
min-public-methods = 1

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
